moved lib tests into seperate file
